+++ mobile/examples/swift/async_await/ContentView.swift
@@ -72,12 +72,8 @@ private extension StreamClient {
         let stream = self
             .newStreamPrototype()
             .setOnResponseHeaders { headers, _, _ in
-                let allHeaders = headers.allHeaders()
-
-                if allHeaders[":status"]?.first == "200",
-                   // TODO(jpsim): Expose an API that enforces case-insensitive lookups
-                   let contentLengthValue = allHeaders["Content-Length"] ??
-                                            allHeaders["content-length"],
+                if headers.value(forName: ":status")?.first == "200",
+                   let contentLengthValue = headers.value(forName: "content-length"),
                    let firstContentLength = contentLengthValue.first,
                    let contentLengthInt = Int64(firstContentLength)
                 {
@@ -87,7 +83,7 @@ private extension StreamClient {
                     return
                 }
 
-                let headerMessage = allHeaders
+                let headerMessage = headers.caseSensitiveHeaders()
                     .map { "\($0.key): \($0.value.joined(separator: ", "))" }
                     .joined(separator: "\n")