+++ mobile/examples/swift/hello_world/ViewController.swift
@@ -70,7 +70,7 @@ final class ViewController: UITableViewController {
         let statusCode = headers.httpStatus.map(String.init) ?? "nil"
         let message = "received headers with status \(statusCode)"
 
-        let headerMessage = headers.allHeaders()
+        let headerMessage = headers.caseSensitiveHeaders()
           .filter { kFilteredHeaders.contains($0.key) }
           .map { "\($0.key): \($0.value.joined(separator: ", "))" }
           .joined(separator: "\n")