+++ mobile/library/swift/RequestTrailersBuilder.swift
@@ -5,13 +5,13 @@ import Foundation
 public final class RequestTrailersBuilder: HeadersBuilder {
   /// Initialize a new instance of the builder.
   public override convenience init() {
-    self.init(headers: [:])
+    self.init(container: HeadersContainer(headers: [:]))
   }
 
   /// Build the request trailers using the current builder.
   ///
   /// - returns: New instance of request trailers.
   public func build() -> RequestTrailers {
-    return RequestTrailers(headers: self.headers)
+    return RequestTrailers(container: self.container)
   }
 }