+++ mobile/docs/root/intro/version_history.rst
@@ -24,6 +24,7 @@ Bugfixes:
 - Android: update Kotlin standard libraries to 1.6.21 (:issue:`#2256 <2256>`)
 - fix bug where finalStreamIntel was not consistently set on cancel (:issue:`#2285 <2285>`)
 - iOS: fix termination crash in ProvisionalDispatcher (:issue:`#2059 <2059>`)
+- iOS: make headers lookup in ``HeadersBuilder`` and ``Headers`` case-insensitive. Rename ``allHeaders`` method to ``caseSensitiveHeaders``. (:issue:`#2383 <2383>`)
 - iOS: use correct DNS resolver when using C++ config builder (:issue: `#2378 <2378 >`)
 
 Features: